refactor: move api files to used component folder
