+++ src/Widget/EnglishCard/fireStore.ts
@@ -1,4 +1,3 @@
-/* eslint-disable import/prefer-default-export */
 import { initializeApp } from 'firebase/app';
 import {
   getFirestore,
