+++ src/Widget/News/News.tsx
@@ -1,7 +1,7 @@
 import React, { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import ReactLoading from 'react-loading';
-import { fetchNews } from '../../utils/backendApis';
+import fetchNews from './newsApis';
 import NewsItem, { NewsData } from './NewsItem';
 import { afterOneHour } from '../../utils/lib';
 import useLocalStorage from '../../hooks/useLocalStorage';
