+++ src/Widget/News/newsApis.ts
@@ -1,7 +1,6 @@
-/* eslint-disable import/prefer-default-export */
 const baseUrl = 'https://tapshaker-server.herokuapp.com/';
 
-export async function fetchNews(countryName: string) {
+export default async function fetchNews(countryName: string) {
   const url = `${baseUrl}news?country=${countryName}`;
   const res = await fetch(url);
   const data = await res.json();
